Added Universal Listener class to project
- Added class to project (not used yet)
- Removed all references to my static logger
- Commented out an odd connect statement

Since this class was developed back in the days of Qt5, and also
developed strictly with Linux and the gcc compiler in mind, I've had to
remove references to my static logger class, which required that
specific platform and compiler to output the class and method log lines
were called from. Also, I won't be creating a log file for this project.

I suspect the use of `of` in the connect statement was related to Qt5,
but I'm not familiar with it (my past self probably was). I'll have to
look into that a little more.