Commented out text anchor attempt
Attempted to anchor vertical text.
Was not working to center it vertially inside a row parent, as it has no
anchor point.
Just noticed a runtime error reported where we can't anchor it to an
explicit id that is not the parent.

This is almost html level centering. Why is this problem so hard in
programming?