Applied outputtting to a signal that happens much earlier
We get the printout immediately this way.

Before, the listener was waiting for more data. This is because I
sometimes stream very large files to my TCP listeners. I don't want to
parse until I have everything, that means waiting for a timeout.

But in this case, I want to just print bytes as I get them.